Fix dialog variant test cleanup
