+++ .changeset/silver-dialog-tests.md
@@ -1,0 +1,5 @@
+---
+"kumo-svelte": patch
+---
+
+Stabilize dialog variant exports so fixed-width dialog sizing can be tested without mounting dialog overlays.
