+++ packages/kumo-svelte/src/lib/components/dialog/index.ts
@@ -1,1 +1,2 @@
 export { default as Dialog } from './Dialog.svelte';
+export { KUMO_DIALOG_DEFAULT_VARIANTS, KUMO_DIALOG_VARIANTS } from './Dialog.variants';
